+++ Tests/MapboxNavigationPackageTests/URLDataCacheTests.swift
@@ -4,6 +4,8 @@ import XCTest
 
 class URLDataCacheTest: TestCase {
     let url = ShieldImage.i280.baseURL
+    let secondURL = URL(string: "https://www.mapbox.com/second-shield.png")!
+    let data = Fixture.JSONFromFileNamed(name: "sprite-info")
     var cache: URLDataCache!
 
     private static var cacheURL: URL {
@@ -26,69 +28,55 @@ class URLDataCacheTest: TestCase {
         super.tearDown()
     }
 
-    private func exampleResponse(with storagePolicy: URLCache.StoragePolicy) -> CachedURLResponse {
-        let data = Fixture.JSONFromFileNamed(name: "sprite-info")
-        let response = URLResponse(url: url, mimeType: nil, expectedContentLength: data.count, textEncodingName: nil)
-        let cachedResponse = CachedURLResponse(response: response, data: data, storagePolicy: storagePolicy)
-        return cachedResponse
-    }
-
     func testStoreCache() {
-        let response = exampleResponse(with: .allowed)
+        cache.store(data, for: url)
 
-        cache.store(response, for: url)
-
-        let cachedResponse = cache.response(for: url)
-        XCTAssertNotNil(cachedResponse)
-        XCTAssertEqual(cachedResponse, response)
+        let cachedData = cache.data(for: url)
+        XCTAssertNotNil(cachedData)
+        XCTAssertEqual(cachedData, data)
+        XCTAssertGreaterThan(cache.currentMemoryUsage, 0)
     }
 
     func testClearCache() {
-        let response = exampleResponse(with: .allowed)
-
-        cache.store(response, for: url)
-        XCTAssertEqual(cache.response(for: url), response)
+        cache.store(data, for: url)
+        XCTAssertEqual(cache.data(for: url), data)
 
         cache.clearCache()
-        XCTAssertNil(cache.response(for: url)?.data)
+        XCTAssertNil(cache.data(for: url))
         XCTAssertEqual(cache.currentMemoryUsage, 0)
+        XCTAssertEqual(cache.currentDiskUsage, 0)
     }
 
     func testRemoveRequestCache() {
-        let response = exampleResponse(with: .allowed)
-
-        cache.store(response, for: url)
-        XCTAssertNotNil(cache.response(for: url))
+        cache.store(data, for: url)
+        cache.store(data, for: secondURL)
+        XCTAssertNotNil(cache.data(for: url))
+        XCTAssertNotNil(cache.data(for: secondURL))
 
         cache.removeCache(for: url)
-        XCTAssertNil(cache.response(for: url)?.data)
+        XCTAssertNil(cache.data(for: url))
+        XCTAssertEqual(cache.data(for: secondURL), data)
     }
 
     func testStoreCacheInMemoryOnly() {
-        let response = exampleResponse(with: .allowedInMemoryOnly)
+        cache.store(data, for: url)
 
-        cache.store(response, for: url)
-
-        let cachedResponse = cache.response(for: url)
-        XCTAssertEqual(cachedResponse, response)
-        XCTAssertEqual(cache.currentMemoryUsage, response.data.count)
+        let cachedData = cache.data(for: url)
+        XCTAssertEqual(cachedData, data)
+        XCTAssertEqual(cache.currentMemoryUsage, data.count)
     }
 
     func testStoreCacheWithMemoryWarning() {
-        let response = exampleResponse(with: .allowed)
-
-        cache.store(response, for: url)
-        XCTAssertEqual(cache.response(for: url), response)
+        cache.store(data, for: url)
+        XCTAssertEqual(cache.data(for: url), data)
 
         NotificationCenter.default.post(name: UIApplication.didReceiveMemoryWarningNotification, object: nil)
-        XCTAssertEqual(cache.response(for: url), response)
+        XCTAssertEqual(cache.data(for: url), data)
     }
 
     func testStoreCacheOutOfCapacity() {
-        let response = exampleResponse(with: .allowedInMemoryOnly)
-
         let limitCapacity = 1
-        XCTAssertTrue(response.data.count > limitCapacity)
+        XCTAssertTrue(data.count > limitCapacity)
 
         let limitCache = URLDataCache(
             memoryCapacity: limitCapacity,
@@ -97,9 +85,19 @@ class URLDataCacheTest: TestCase {
         )
         limitCache.clearCache()
 
-        limitCache.store(response, for: url)
-        XCTAssertNil(cache.response(for: url))
+        limitCache.store(data, for: url)
+        XCTAssertNil(limitCache.data(for: url))
         XCTAssertEqual(limitCache.currentMemoryUsage, 0)
-        // not checking disk usage as it is always non-zero
+        XCTAssertEqual(limitCache.currentDiskUsage, 0)
+    }
+
+    func testStoreCacheOnDisk() {
+        let diskCache = URLDataCache(diskCapacity: data.count * 2, diskCacheURL: Self.cacheURL)
+
+        diskCache.store(data, for: url)
+        XCTAssertGreaterThan(diskCache.currentDiskUsage, 0)
+        cache = URLDataCache(diskCapacity: 0, diskCacheURL: Self.cacheURL)
+
+        XCTAssertEqual(cache.data(for: url), data)
     }
 }